pyproject.toml
[tool.poetry]
name = "neoclient"
version = "0.1.0"
description = ""
authors = ["neeospace <engineering@neospace.ai>"]
license = "Apache-2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "2.32.3"
typing-extensions = "4.12.2"
python-dotenv = "^1.0.1"
pydantic = "^2.10.0"

[tool.poetry.dev-dependencies]
pytest = "8.3.3"
pytest-mock = "3.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
